+++ src/js/likes.js
@@ -1,5 +1,4 @@
-const likeUrl =
-  'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/0tY3ZECsTgObPiElFJSy/likes';
+const likeUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/0tY3ZECsTgObPiElFJSy/likes';
 
 function updateLikeForDOM(name) {
   const itemNames = document.querySelectorAll('.name');
@@ -26,6 +25,7 @@ export const addLike = async (name) => {
   });
 
   updateLikeForDOM(name);
+  return (await postingLike).json();
 };
 
 export const getLike = async () => {
