+++ src/sass/_comment.scss
@@ -1,103 +1,23 @@
 $secondry-color: #f9002b;
 
-body {
-  background-color: black;
-  color: #858fa2;
+.description {
+  position: absolute;
+  transform: translateY(-200px) scale(0);
+  padding: 43px 10px 0;
+  height: 0;
+  top: 0;
+  width: 100%;
+  background-color: rgba(0, 0, 0, 0.767);
 
-  header {
-    .logo {
-      color: $primary;
-      font-size: 40px;
-      padding: 20px 0;
-      :hover {
-        color: $secondry-color;
-        transition: all 0.5s ease-in-out;
-      }
-    }
-
-    h1 {
-      font-family: 'Tourney', cursive;
-      font-weight: 400;
-      margin-bottom: 0;
-      background-color: $secondry-color;
-      background-image: linear-gradient(45deg, #f9002b, #1111f3);
-      background-size: 100%;
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-      &:hover {
-        cursor: pointer;
-        transition: all 0.8s ease-in-out;
-        background-image: linear-gradient(45deg, #1111f3, #f9002b);
-      }
-    }
+  h6 {
+    font-size: 15px;
+    font-weight: bold;
+    padding: 0;
+    margin: 0;
   }
 
-  main {
-    padding: 100px 50px;
-    .img {
-      width: 300px;
-      height: 400px;
-    }
-
-    #list {
-      display: grid;
-      grid-template-columns: repeat(3, 1fr);
-      gap: 80px 30px;
-
-      li {
-        // border: 1px solid white;
-        position: relative;
-        .comment {
-          border: 1px solid #f9002b;
-          background: transparent;
-          color: #ffff;
-          font-size: 0.8rem;
-          padding: 6px 18px;
-          font-weight: 600;
-          border-radius: 3px;
-          i {
-            color: $secondry-color;
-            font-size: 20px;
-            margin-right: 5px;
-          }
-
-          &:hover {
-            background-color: $primary;
-            border: 1px solid $primary;
-          }
-        }
-        .title {
-          p {
-            padding: 0;
-            margin: 0;
-          }
-          h2 {
-            font-size: 1.15rem;
-            padding: 0;
-            margin: 0;
-          }
-
-          a {
-            text-decoration: none;
-            &:hover {
-              cursor: pointer;
-              transition: all 0.2s ease;
-              color: $secondry-color;
-            }
-          }
-        }
-        &:hover {
-          transform: scale(1.09);
-          cursor: pointer;
-          transition: all 0.5s ease-in-out;
-          .description {
-            transition: all 0.5s ease-in-out;
-            height: auto;
-            transform: translate(0) scale(1);
-          }
-        }
-      }
-    }
+  p {
+    font-size: 13px;
   }
 }
 
@@ -105,10 +25,10 @@ body {
   height: 100vh;
   position: fixed;
   top: 0;
-  width: 100%;
+  width: 101%;
   overflow-y: scroll;
   padding: 100px;
-  background-color: rgba(0, 0, 0, 0.938);
+  background-color: rgba(0, 0, 0, 0.849);
 
   .pop-img {
     width: 60%;
@@ -119,14 +39,17 @@ body {
     .sam-p {
       width: 40%;
       font-size: 0.9rem;
+
       h6 {
         font-weight: bold;
       }
+
       p {
         margin-left: 15px;
       }
     }
   }
+
   h2 {
     color: rgb(231, 231, 231);
     font-weight: bold;
@@ -144,11 +67,13 @@ body {
       margin: 10px;
       width: 60%;
       padding: 15px 10px;
+
       &:focus {
-        outline: 1px solid $secondry-color;
+        outline: 2px solid $primary;
         color: white;
       }
     }
+
     .input-name {
       line-height: 18px;
     }
@@ -164,11 +89,13 @@ body {
       font-weight: 600;
       border-radius: 3px;
       margin: 20px 5px;
+
       &:hover {
         background-color: transparent;
         border: 1px solid $secondry-color;
         transition: all 0.4s ease-in-out;
       }
+
       i {
         position: inherit;
         color: $secondry-color;
@@ -185,9 +112,123 @@ body {
     top: 0;
     right: 0;
     z-index: 999;
+
     &:hover {
       color: $secondry-color;
       cursor: pointer;
+    }
+  }
+}
+
+body {
+  background-color: black;
+  color: #858fa2;
+
+  header {
+    .logo {
+      color: $primary;
+      font-size: 40px;
+      padding: 20px 0;
+
+      :hover {
+        color: $secondry-color;
+        transition: all 0.5s ease-in-out;
+      }
+    }
+
+    h1 {
+      font-family: 'Tourney', cursive;
+      font-weight: 400;
+      margin-bottom: 0;
+      background-color: $secondry-color;
+      background-image: linear-gradient(45deg, #f9002b, #1111f3);
+      background-size: 100%;
+      -webkit-background-clip: text;
+      -webkit-text-fill-color: transparent;
+
+      &:hover {
+        cursor: pointer;
+        transition: all 0.8s ease-in-out;
+        background-image: linear-gradient(45deg, #1111f3, #f9002b);
+      }
+    }
+  }
+
+  main {
+    padding: 100px 50px;
+
+    .img {
+      width: 300px;
+      height: 400px;
+    }
+
+    #list {
+      display: grid;
+      grid-template-columns: repeat(3, 1fr);
+      gap: 80px 30px;
+
+      li {
+        // border: 1px solid white;
+        position: relative;
+        padding: 10px;
+
+        .comment {
+          border: 1px solid #f9002b;
+          background: transparent;
+          color: #ffff;
+          font-size: 0.8rem;
+          padding: 6px 18px;
+          font-weight: 600;
+          border-radius: 3px;
+
+          i {
+            color: $secondry-color;
+            font-size: 20px;
+            margin-right: 5px;
+          }
+
+          &:hover {
+            background-color: $primary;
+            border: 1px solid $primary;
+          }
+        }
+
+        .title {
+          p {
+            padding: 0;
+            margin: 0;
+          }
+
+          h2 {
+            font-size: 1.15rem;
+            padding: 0;
+            margin: 0;
+          }
+
+          a {
+            text-decoration: none;
+
+            &:hover {
+              cursor: pointer;
+              transition: all 0.2s ease;
+              color: $secondry-color;
+            }
+          }
+        }
+
+        &:hover {
+          transform: scale(1.09);
+          box-shadow: 1px 1px 5px 2px #f9002934;
+          cursor: pointer;
+          transition: all 0.5s ease-in-out;
+
+          .description {
+            transition: all 0.5s ease-in-out;
+            height: auto;
+            transform: translate(0) scale(1);
+          }
+        }
+      }
     }
   }
 }
@@ -199,7 +240,8 @@ body {
   position: absolute;
   top: 0;
   width: 100%;
-  i {
+
+  .add-icon {
     font-size: 20px;
   }
 }
@@ -208,21 +250,11 @@ footer {
   background-color: #f9002946;
 }
 
-.description {
-  position: absolute;
-  transform: translateY(-200px) scale(0);
-  padding: 43px 10px 0;
-  height: 0;
-  top: 0;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.767);
-  h6 {
-    font-size: 15px;
-    font-weight: bold;
-    padding: 0;
-    margin: 0;
-  }
-  p {
-    font-size: 13px;
-  }
+.pop-html {
+  overflow-y: hidden;
+}
+
+.pop-html main ul,
+.pop-html header {
+  filter: blur(20px);
 }
